apps/backend-api/pyproject.toml
[project]
name = "cxg-backend-api"
version = "0.1.0"
description = "CXG Platform API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    
    # SQLAlchemy (2.0+)
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",              # PostgreSQL async driver
    "psycopg2-binary>=2.9.9",       # PostgreSQL sync driver (fallback)
    "greenlet>=3.0.0",              # SQLAlchemy async 필수
    
    # Pydantic (FastAPI와 호환)
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database migrations
    "alembic>=1.13.0",
    
    # Redis
    "redis>=5.0.0",
    
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",  # Pin to 4.x (5.x has strict password length checking)
    "python-multipart>=0.0.6",
    
    # Utilities
    "python-dotenv>=1.0.0",
    
    # GraphQL
    "strawberry-graphql[fastapi]>=0.235.0",
    "aiodataloader>=0.4.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "faker>=22.0.0",
    
    # Code quality (Ruff + mypy 조합)
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    
    # Type stubs
    "types-passlib>=1.7.0",
    "types-python-jose>=3.3.0",
]

ai = [
    "openai>=1.10.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.20",
    "pinecone-client>=3.0.0",
    "tiktoken>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

# ==============================================================================
# Ruff - 린팅 + 포매팅 + import 정렬 (올인원)
# ==============================================================================
[tool.ruff]
# Python 버전
target-version = "py311"

# 줄 길이
line-length = 100

# 제외할 디렉토리
exclude = [
    ".git",
    ".venv",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "__pycache__",
    "alembic/versions",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# 활성화할 규칙 세트
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import 정렬)
    "B",      # flake8-bugbear (버그 가능성)
    "C4",     # flake8-comprehensions (comprehension 개선)
    "UP",     # pyupgrade (최신 Python 문법)
    "ARG",    # flake8-unused-arguments (미사용 인자)
    "SIM",    # flake8-simplify (코드 단순화)
    "TCH",    # flake8-type-checking (타입 체크 import)
    "N",      # pep8-naming (네이밍 컨벤션)
    "ASYNC",  # flake8-async (async 코드 체크)
]

# 무시할 규칙
ignore = [
    "E501",   # line too long (formatter가 처리)
    "B008",   # function calls in argument defaults (FastAPI Depends)
    "C901",   # too complex (필요시 복잡도 허용)
    "ARG001", # unused function argument (FastAPI 의존성)
    "ARG002", # unused method argument
]

# 자동 수정 가능한 규칙
fixable = ["ALL"]
unfixable = []

# 파일별 예외 규칙
[tool.ruff.lint.per-file-ignores]
# 테스트 파일은 느슨하게
"tests/**/*.py" = [
    "ARG",    # unused arguments (fixture 사용)
    "S101",   # assert 사용 허용
]
# 마이그레이션 파일
"alembic/versions/*.py" = [
    "ALL",    # 모든 규칙 무시
]
# __init__.py
"__init__.py" = [
    "F401",   # unused imports (re-export용)
    "F403",   # wildcard imports (re-export용)
    "F405",   # undefined names from wildcard
]

# Import 정렬 (isort 호환)
[tool.ruff.lint.isort]
known-first-party = ["src"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
lines-after-imports = 2

# 포매팅 (Black 호환)
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

# ==============================================================================
# mypy - 타입 체크
# ==============================================================================
[tool.mypy]
python_version = "3.11"
mypy_path = "$MYPY_CONFIG_FILE_DIR/src"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_unreachable = true
strict_equality = true
plugins = ["pydantic.mypy"]

# 모듈별 설정
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_errors = true

# 서드파티 라이브러리 타입 스텁 없을 때
[[tool.mypy.overrides]]
module = [
    "passlib.*",
    "jose.*",
    "asyncpg.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
# models/__init__.py는 re-export 목적
